Add option --root-group
If a gitlab user has access to a large amount of groups, it may take
much time to load an entire tree.  The option --root-group lets the user
specify an exact group and just use it as the root instead of waiting
for the entire tree to be loaded and filtered.